feat: op-sepolia support

+++ scripts/init-bedrock.sh
@@ -7,7 +7,7 @@ source ./scripts/utils.sh
 # Common variables.
 INITIALIZED_FLAG=/shared/initialized.txt
 BEDROCK_JWT_PATH=/shared/jwt.txt
-GETH_DATA_DIR=/geth
+GETH_DATA_DIR=$BEDROCK_DATADIR
 TORRENTS_DIR=/torrents/$NETWORK_NAME
 BEDROCK_TAR_PATH=/downloads/bedrock.tar
 BEDROCK_TMP_PATH=/bedrock-tmp
@@ -29,22 +29,24 @@ elif [ "$NETWORK_NAME" = "op-goerli" ]; then
   BEDROCK_TAR_DOWNLOAD="https://datadirs.optimism.io/goerli-bedrock.tar.zst"
 fi
 
-if [[ "$BEDROCK_TAR_DOWNLOAD" == *.zst ]]; then
-  BEDROCK_TAR_PATH+=".zst"
+if [ -n "${BEDROCK_TAR_DOWNLOAD+x}" ]; then
+  if [[ "$BEDROCK_TAR_DOWNLOAD" == *.zst ]]; then
+    BEDROCK_TAR_PATH+=".zst"
+  fi
+
+  echo "Downloading bedrock.tar..."
+  download $BEDROCK_TAR_DOWNLOAD $BEDROCK_TAR_PATH
+
+  echo "Extracting bedrock.tar..."
+  if [[ "$BEDROCK_TAR_DOWNLOAD" == *.zst ]]; then
+    extractzst $BEDROCK_TAR_PATH $GETH_DATA_DIR
+  else
+    extract $BEDROCK_TAR_PATH $GETH_DATA_DIR
+  fi
+
+  # Remove tar file to save disk space
+  rm $BEDROCK_TAR_PATH
 fi
-
-echo "Downloading bedrock.tar..."
-download $BEDROCK_TAR_DOWNLOAD $BEDROCK_TAR_PATH
-
-echo "Extracting bedrock.tar..."
-if [[ "$BEDROCK_TAR_DOWNLOAD" == *.zst ]]; then
-  extractzst $BEDROCK_TAR_PATH $GETH_DATA_DIR
-else
-  extract $BEDROCK_TAR_PATH $GETH_DATA_DIR
-fi
-
-# Remove tar file to save disk space
-rm $BEDROCK_TAR_PATH
 
 echo "Creating JWT..."
 mkdir -p $(dirname $BEDROCK_JWT_PATH)
